Remove Android code from /ports/servo/
