+++ ports/servo/Cargo.toml
@@ -33,7 +33,7 @@ googlevr = ["libservo/googlevr"]
 oculusvr = ["libservo/oculusvr"]
 unstable = ["libservo/unstable"]
 
-[dependencies]
+[target.'cfg(not(target_os = "android"))'.dependencies]
 backtrace = "0.3"
 bitflags = "1.0"
 euclid = "0.17"
@@ -44,8 +44,6 @@ libservo = {path = "../../components/servo"}
 log = "0.4"
 tinyfiledialogs = "3.0"
 winit = {version = "0.16", features = ["icon_loading"]}
-
-[target.'cfg(not(target_os = "android"))'.dependencies]
 sig = "0.1"
 
 [target.'cfg(any(target_os = "linux", target_os = "macos"))'.dependencies]
@@ -53,10 +51,6 @@ osmesa-sys = "0.1.2"
 
 [target.'cfg(target_os = "linux")'.dependencies]
 x11 = "2.0.0"
-
-[target.'cfg(target_os = "android")'.dependencies]
-android_injected_glue = "0.2"
-servo-egl = "0.2"
 
 [target.'cfg(target_os = "windows")'.dependencies]
 winapi = "0.2"
